Nearly finished game gameLogic
changeTurn function adds a player token 'X' or 'O' to a square on the grid upon clicking.
changeTurn function is also responsible for checking to see if the tile is a valid tile and changes player turns.
pushToArray function adds 'X' or 'O' to the board array
winCondition function checks for occupied tiles and win condition based on board array index